requirements.txt
fastapi
uvicorn
jinja2
python-multipart
google-genai
pdfplumber
python-docx
speechrecognition
pydub
